# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Project Alpha API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /upload/incorporation-cert:
    post:
      operationId: uploadIncorporationCert
      tags: [storage]
      summary: Upload a Certificate of Incorporation
      description: Accepts a multipart PDF or image file. Stores it in object storage and returns the serving URL. Used during service provider signup.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF or image file (max 10MB)
              required:
                - file
      responses:
        "201":
          description: File uploaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadCertResponse"
        "400":
          description: No file or invalid file type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Upload failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UploadCertResponse:
      type: object
      properties:
        objectPath:
          type: string
          description: Normalized object path (e.g. /objects/uploads/uuid)
        fileUrl:
          type: string
          description: Auth-gated serving URL for the uploaded file (/api/storage/objects/...). Requires a valid Authorization header. Not a public GCS URL.
        metadata:
          type: object
          properties:
            name:
              type: string
            size:
              type: integer
            contentType:
              type: string
      required:
        - objectPath
        - fileUrl
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required:
        - error